+++ extensions/ovh/src/types.ts
@@ -50,23 +50,30 @@ export type DNSRecord = {
   zone: string;
 };
 export type DomainServiceInfo = {
-  contactAdmin: string
-  renewalType: 'automaticForcedProduct' | 'automaticV2012' | 'automaticV2014' | 'automaticV2016' | 'manual' | 'oneShot' | 'option',
-  expiration: string
-  canDeleteAtExpiration: boolean,
-  possibleRenewPeriod: [ number ],
-  contactBilling: string
-  serviceId: number,
+  contactAdmin: string;
+  renewalType:
+    | "automaticForcedProduct"
+    | "automaticV2012"
+    | "automaticV2014"
+    | "automaticV2016"
+    | "manual"
+    | "oneShot"
+    | "option";
+  expiration: string;
+  canDeleteAtExpiration: boolean;
+  possibleRenewPeriod: [number];
+  contactBilling: string;
+  serviceId: number;
   domain: string;
-  engagedUpTo: string | null,
-  creation: string
+  engagedUpTo: string | null;
+  creation: string;
   renew: {
-    deleteAtExpiration: boolean
-    forced: boolean
-    manualPayment: boolean | null
-    period: number | null
-    automatic: boolean
-  } | null,
-  status: 'autorenewInProgress' | 'expired' | 'inCreation' | 'ok' | 'pendingDebt' | 'unPaid',
-  contactTech: string
-}
+    deleteAtExpiration: boolean;
+    forced: boolean;
+    manualPayment: boolean | null;
+    period: number | null;
+    automatic: boolean;
+  } | null;
+  status: "autorenewInProgress" | "expired" | "inCreation" | "ok" | "pendingDebt" | "unPaid";
+  contactTech: string;
+};