+++ extensions/ovh/CHANGELOG.md
@@ -1,5 +1,10 @@
 # OVHcloud Changelog
 
+## [Update Domain & DNS Service Information] - {PR_MERGE_DATE}
+
+- Update Domain Service Information (renewal)
+- Update DNS Zone Service Information (renewal)
+
 ## [Manage Domains] - 2025-10-10
 
 - Manage Domains, DNS Records and Nameservers